revert previous config
